code reorg; starboard minor improvements
